scripts/requirements.txt
Pillow
poeditor
python-dotenv
requests
